refact(GQLV2): Move root mutations to their own file
