+++ server/graphql/v2/mutation/AccountMutations.ts
@@ -14,15 +14,11 @@ import { cloneDeep, set } from 'lodash';
 
 import { types as COLLECTIVE_TYPE } from '../../../constants/collectives';
 import plans from '../../../constants/plans';
-import cache, { purgeAllCachesForAccount, purgeGQLCacheForCollective } from '../../../lib/cache';
-import { purgeCacheForPage } from '../../../lib/cloudflare';
-import { invalidateContributorsCache } from '../../../lib/contributors';
+import cache from '../../../lib/cache';
 import { crypto } from '../../../lib/encryption';
-import { mergeAccounts, simulateMergeAccounts } from '../../../lib/merge-accounts';
 import { verifyTwoFactorAuthenticatorCode } from '../../../lib/two-factor-authentication';
 import models, { sequelize } from '../../../models';
 import { Forbidden, NotFound, Unauthorized, ValidationFailed } from '../../errors';
-import { AccountCacheType } from '../enum/AccountCacheType';
 import { AccountTypeToModelMapping } from '../enum/AccountType';
 import { Policy } from '../enum/Policy';
 import { idDecode } from '../identifiers';
@@ -31,7 +27,6 @@ import { AccountUpdateInput } from '../input/AccountUpdateInput';
 import { Account } from '../interface/Account';
 import { Host } from '../object/Host';
 import { Individual } from '../object/Individual';
-import { MergeAccountsResponse } from '../object/MergeAccountsResponse';
 import AccountSettingsKey from '../scalar/AccountSettingsKey';
 
 const AddTwoFactorAuthTokenToIndividualResponse = new GraphQLObjectType({
@@ -392,79 +387,6 @@ const accountMutations = {
       }
 
       return account;
-    },
-  },
-  clearCacheForAccount: {
-    type: new GraphQLNonNull(Account),
-    description: '[Root only] Clears the cache for a given account',
-    args: {
-      account: {
-        type: new GraphQLNonNull(AccountReferenceInput),
-        description: 'Account to clear the cache for',
-      },
-      type: {
-        type: new GraphQLNonNull(new GraphQLList(AccountCacheType)),
-        description: 'Types of cache to clear',
-        defaultValue: ['CLOUDFLARE', 'GRAPHQL_QUERIES', 'CONTRIBUTORS'],
-      },
-    },
-    async resolve(_: void, args, req: express.Request): Promise<Record<string, unknown>> {
-      if (!req.remoteUser?.isRoot()) {
-        throw new Forbidden('Only root users can perform this action');
-      }
-
-      const account = await fetchAccountWithReference(args.account, { throwIfMissing: true });
-
-      if (args.type.includes('CLOUDFLARE')) {
-        purgeCacheForPage(`/${account.slug}`);
-      }
-      if (args.type.includes('GRAPHQL_QUERIES')) {
-        purgeGQLCacheForCollective(account.slug);
-      }
-      if (args.type.includes('CONTRIBUTORS')) {
-        await invalidateContributorsCache(account.id);
-      }
-
-      return account;
-    },
-  },
-  mergeAccounts: {
-    type: new GraphQLNonNull(MergeAccountsResponse),
-    description: '[Root only] Merge two accounts, returns the result account',
-    args: {
-      fromAccount: {
-        type: new GraphQLNonNull(AccountReferenceInput),
-        description: 'Account to merge from',
-      },
-      toAccount: {
-        type: new GraphQLNonNull(AccountReferenceInput),
-        description: 'Account to merge to',
-      },
-      dryRun: {
-        type: new GraphQLNonNull(GraphQLBoolean),
-        description: 'If true, the result will be simulated and summarized in the response message',
-        defaultValue: true,
-      },
-    },
-    async resolve(_: void, args, req: express.Request): Promise<Record<string, unknown>> {
-      if (!req.remoteUser?.isRoot()) {
-        throw new Forbidden('Only root users can perform this action');
-      }
-
-      const fromAccount = await fetchAccountWithReference(args.fromAccount, { throwIfMissing: true });
-      const toAccount = await fetchAccountWithReference(args.toAccount, { throwIfMissing: true });
-
-      if (args.dryRun) {
-        const message = await simulateMergeAccounts(fromAccount, toAccount);
-        return { account: toAccount, message };
-      } else {
-        const warnings = await mergeAccounts(fromAccount, toAccount, req.remoteUser.id);
-        await Promise.all([purgeAllCachesForAccount(fromAccount), purgeAllCachesForAccount(toAccount)]).catch(() => {
-          // Ignore errors
-        });
-        const message = warnings.join('\n');
-        return { account: await toAccount.reload(), message: message || null };
-      }
     },
   },
   setPolicies: {
